Address review round 2: annotate ITestFilterProviderAttribute.FilterType
The adapter now reads the filter type through ITestFilterProviderAttribute and
passes it to InstantiateTestFilter, whose parameter requires
[DynamicallyAccessedMembers(PublicParameterlessConstructor)]. Trimming resolves
annotations from the declaring member, so leaving the interface member
unannotated broke the annotation flow that the concrete non-generic property
already had, reported as IL2093:

  dotnet build ... -p:EnableTrimAnalyzer=true -p:EnableAotAnalyzer=true
  warning IL2093: 'DynamicallyAccessedMemberTypes' ... on the return value of
  '...TestFilterProviderAttribute.FilterType.get' don't match overridden return
  value of '...ITestFilterProviderAttribute.FilterType.get'.

Annotate the interface member and the generic attribute's property so all three
declarations match; that build is now clean of IL2093.

Also broaden the acceptance coverage. The interface-based lookup replaced the
lookup used by the already-shipped non-generic attribute, so the risk is not
limited to the new generic form. The asset now multi-targets and selects the
shape with #if NET, which exercises the generic form on .NET and the non-generic
form on .NET Framework -- the previously untested leg -- and doubles as the
worked example of the multi-targeting pattern, which is now documented on the
attribute itself.

Copilot-Session: 0a7d9aa7-3a75-4e37-8f05-dee535d69648
